Logo - Versions 3 and 4
This is the third and fourth version of the logo. I took a different route compared to the first two. I wanted to make a logo that has a relation to do with weights. In version 3 it's much more compact with the shadow highlights of the font where as in version 4 it's longer and made to look more seamless.