Logo - Version 5
This is the fifth version of the logo. I went for a clean design that's very minimalistic but with a very loose and funky twist. I drew a carrot in a simple way because I personally believe a simple logo is much more professional and outstanding compared to the crazy and bizarre "there's a lot going on" type of logos.