Logo - Version 6
This is the fifth version of the logo. I went for a clean design that's very minimalistic but with a very loose and funky twist. I used a carrots colour theme to give the logo a creative and unique effect. The logo gives a much bigger vibrance and adds a lot more colour to the overall theme.